Refresh version component on replacing diagram
